Refactor create-org command
Currently, there are two patterns that commands in this repo follow,
the "old" pattern and the "new" pattern. The old pattern is generally
considered worse, so we are refactoring the create org command to follow
the new pattern. The behavior should remain unchanged, but the code
should be easier to maintain and should be more idiomatic go.

Co-authored-by: Tom Viehman <tviehman@pivotal.io>